Adding active row highlight
It would be nice to have an easy way to quickly see which charge code you're looking at later in the week if you have a ton of charge codes.

Added a hover effect to highlight your current row.  I'm sure you can get this looking a little slicker, but here's an idea to start with.